+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 
 ### Added
+- fix(debian-linux): 🔧 update firefox debian package ppa script
 - git: ssh key signing
 - shell: aliases for editing shell history with fd
 - git: aliases for git clone `git clone p:repo-name`